// src/app/api/model/rapportDto.ts
/**
 * OpenApi specification - Ousmane Vincent Dione
 * OpenApi documentation for Stock Management
 *
 * The version of the OpenAPI document: 1.0
 * Contact: dioneousmanevincent@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EvolutionJournaliereDto } from './evolutionJournaliereDto';
import { EvolutionMensuelleDto } from './evolutionMensuelleDto';
import { StatutCommandesDto } from './statutCommandesDto';
import { ArticleQuantiteDto } from './articleQuantiteDto';
import { ArticleVenteDto } from './articleVenteDto';


export interface RapportDto { 
    dateDebut?: string;
    dateFin?: string;
    typeRapport?: string;
    chiffreAffaireTotal?: number;
    nombreVentes?: number;
    nombreCommandesClients?: number;
    nombreCommandesFournisseurs?: number;
    montantMoyenVente?: number;
    evolutionJournaliere?: Array<EvolutionJournaliereDto>;
    evolutionMensuelle?: Array<EvolutionMensuelleDto>;
    articlesPlusVendus?: Array<ArticleVenteDto>;
    articlesQuantite?: Array<ArticleQuantiteDto>;
    statutCommandesClients?: StatutCommandesDto;
    statutCommandesFournisseurs?: StatutCommandesDto;
}