// src/app/api/model/rapportSearchCriteria.ts
/**
 * OpenApi specification - Ousmane Vincent Dione
 * OpenApi documentation for Stock Management
 *
 * The version of the OpenAPI document: 1.0
 * Contact: dioneousmanevincent@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RapportSearchCriteria { 
    periode?: RapportSearchCriteria.PeriodeEnum;
    dateDebut?: string;
    dateFin?: string;
    avecDetails?: boolean;
    limitArticles?: number;
    idEntreprise?: number;
}
export namespace RapportSearchCriteria {
    export type PeriodeEnum = 'HIER' | 'SEMAINE' | 'MOIS_EN_COURS' | 'ANNEE_EN_COURS' | 'PERSONNALISEE';
    export const PeriodeEnum = {
        Hier: 'HIER' as PeriodeEnum,
        Semaine: 'SEMAINE' as PeriodeEnum,
        MoisEnCours: 'MOIS_EN_COURS' as PeriodeEnum,
        AnneeEnCours: 'ANNEE_EN_COURS' as PeriodeEnum,
        Personnalisee: 'PERSONNALISEE' as PeriodeEnum
    };
}